Added a simple captioning mechanism
To add a caption, simply rename the photo. It must starts with a '@', and the rest of the name (excluding the extension) is
used as the caption.

Markdown syntax is accepted.

Note that some characters are forbidden, like the # or the / as they lead to either invalid file names or invalid URLs.
In order to make it possible to input links in the captions, pipes ('|') are replaced by '/'